extract the startup code
