+++ src/main.rs
@@ -1,10 +1,5 @@
-use std::net::TcpListener;
-
-use sqlx::postgres::PgPoolOptions;
-
 use newsletter::configuration::get_configuration;
-use newsletter::email_client::EmailClient;
-use newsletter::startup::run;
+use newsletter::startup::build;
 use newsletter::telemetry::{get_subscriber, init_subscriber};
 
 #[tokio::main]
@@ -13,26 +8,7 @@ async fn main() -> std::io::Result<()> {
     init_subscriber(subscriber);
 
     let configuration = get_configuration().expect("Failed to read configuration.");
-    let connection_pool = PgPoolOptions::new()
-        .connect_timeout(std::time::Duration::from_secs(3))
-        .connect_lazy_with(configuration.database.with_db());
-    let sender_email = configuration
-        .email_client
-        .sender()
-        .expect("Invalid sender email address.");
-    let timeout = configuration.email_client.timeout();
-    let email_client = EmailClient::new(
-        configuration.email_client.base_url,
-        sender_email,
-        configuration.email_client.authorization_token,
-        timeout,
-    );
-    let address = format!(
-        "{}:{}",
-        configuration.application.host, configuration.application.port
-    );
-    let listener = TcpListener::bind(address)?;
-
-    run(listener, connection_pool, email_client)?.await?;
+    let server = build(configuration).await?;
+    server.await?;
     Ok(())
 }
